Create bootstrapper only once per module
Creating the bootstrapper is slow. This changes it from creating it
once per test to just once per module (test class for the module).
Currently all the tests could share the same bootstrapper and shave
off some more time, but this change is just the quick win by swapping
the NUnit attribute.