Introduce Rake and Centroid
Rake is for driving a project's automated tasks and Centroid helps
with managing configuration across languages.

This also adds a rake task called ok that currently checks to see
the local system has all the necessary environment variables. This
could later run tests, or whatever else to help with getting a
system to the status of OK.